requirements.txt
langchain>=0.1.0
langchain-community>=0.3.1
langchain-openai>=0.0.3
openai>=1.6.1
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
tiktoken>=0.5.2